refactor: :recycle: auth response 패키지 구조 변경 및 JsonNaming 어노테이션 적용
