--- be/src/main/java/com/jjikmuk/sikdorak/auth/controller/dto/response/OAuthTokenResponse.java
@@ -1,25 +1,0 @@
-package com.jjikmuk.sikdorak.auth.controller.dto.response;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-import lombok.ToString;
-
-@ToString
-@Getter
-public class OAuthTokenResponse {
-
-    @JsonProperty("token_type")
-    private String tokenType;
-
-    @JsonProperty("access_token")
-    private String accessToken;
-
-    @JsonProperty("expires_in")
-    private int expiresIn;
-
-    @JsonProperty("refresh_token")
-    private String refreshToken;
-
-    @JsonProperty("refresh_token_expires_in")
-    private int refreshTokenExpiresIn;
-}
