+++ be/src/main/java/com/jjikmuk/sikdorak/auth/controller/response/OAuthTokenResponse.java
@@ -1,0 +1,20 @@
+package com.jjikmuk.sikdorak.auth.controller.response;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.Getter;
+
+@Getter
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public class OAuthTokenResponse {
+
+    private String tokenType;
+
+    private String accessToken;
+
+    private int expiresIn;
+
+    private String refreshToken;
+
+    private int refreshTokenExpiresIn;
+}
