+++ be/src/main/java/com/jjikmuk/sikdorak/auth/service/OAuthApiClient.java
@@ -1,6 +1,6 @@
 package com.jjikmuk.sikdorak.auth.service;
 
-import com.jjikmuk.sikdorak.auth.controller.dto.response.KakaoAccountResponse;
+import com.jjikmuk.sikdorak.auth.controller.response.KakaoAccountResponse;
 import com.jjikmuk.sikdorak.common.config.feignclient.FeignClientHeaderConfiguration;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
