+++ be/src/main/java/com/jjikmuk/sikdorak/common/config/feignclient/KakaoAccountResponseDeserializer.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.jjikmuk.sikdorak.auth.controller.dto.response.KakaoAccountResponse;
+import com.jjikmuk.sikdorak.auth.controller.response.KakaoAccountResponse;
 
 import java.io.IOException;
